Isolate db state checks to dispatch activity
