+++ app/src/org/commcare/dalvik/application/InitializationHelper.java
@@ -4,41 +4,43 @@ import android.app.Activity;
 
 import org.commcare.android.util.SessionUnavailableException;
 import org.commcare.dalvik.R;
-import org.commcare.dalvik.activities.CommCareHomeActivity;
+import org.commcare.dalvik.activities.DispatchActivity;
 import org.javarosa.core.services.locale.Localization;
 
 /**
  * @author Phillip Mates (pmates@dimagi.com).
  */
 public class InitializationHelper {
-    public static void checkDbState(Activity activity) {
+    public static boolean isDbInBadState(DispatchActivity activity) {
         int dbState = CommCareApplication._().getDatabaseState();
         if (dbState == CommCareApplication.STATE_MIGRATION_FAILED) {
             CommCareApplication._().triggerHandledAppExit(activity,
                     activity.getString(R.string.migration_definite_failure),
                     activity.getString(R.string.migration_failure_title), false);
+            return true;
         } else if (dbState == CommCareApplication.STATE_MIGRATION_QUESTIONABLE) {
             CommCareApplication._().triggerHandledAppExit(activity,
                     activity.getString(R.string.migration_possible_failure),
                     activity.getString(R.string.migration_failure_title), false);
+            return true;
         } else if (dbState == CommCareApplication.STATE_CORRUPTED) {
             handleDamagedApp(activity);
+            return true;
         }
+        return false;
     }
 
-    public static void handleDamagedApp(Activity activity) {
+    public static void handleDamagedApp(DispatchActivity activity) {
         if (!CommCareApplication._().isStorageAvailable()) {
             createNoStorageDialog(activity);
         } else {
             // See if we're logged in. If so, prompt for recovery.
             try {
                 CommCareApplication._().getSession();
-                // TODO PLM, work with dispatch activity
-                activity.showDialog(CommCareHomeActivity.DIALOG_CORRUPTED);
+                activity.showDialog(DispatchActivity.DIALOG_CORRUPTED);
             } catch(SessionUnavailableException e) {
                 // Otherwise, log in first
-                // TODO PLM
-                //launchLogin();
+                activity.launchLogin();
             }
         }
     }
@@ -48,5 +50,4 @@ public class InitializationHelper {
                 Localization.get("app.storage.missing.message"),
                 Localization.get("app.storage.missing.title"));
     }
-
 }
